# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    ApiError:
      type: object
      description: "The error envelope: a message, plus a reason or a code on a 409."
      properties:
        message:
          type: string
          description: "What went wrong, in words. Always present."
        reason:
          $ref: "#/components/schemas/Reason"
          type: string
          description: "Which refusal this is. Present on the 409s the release and\
            \ integrate flows raise, absent on every other error. A value a client\
            \ does not recognise is shown as the message verbatim."
        conflicts:
          type: array
          items:
            type: string
          description: The conflicted paths. Present only for CONFLICT and MERGE_CONFLICT.
        code:
          type: string
          description: "The refusal's own name, where it has one (e.g. RUNNER_OWNS_WORKSPACES).\
            \ Absent otherwise."
        workspaceIds:
          type: array
          items:
            type: integer
            format: int64
          description: The ACTIVE workspace row ids a runner still owns. Present only
            with RUNNER_OWNS_WORKSPACES.
        workspaces:
          type: array
          items:
            $ref: "#/components/schemas/RunnerOwnedWorkspace"
          description: "The ACTIVE workspaces a runner still owns, each with its repository\
            \ and branch so a client can link it. Present only with RUNNER_OWNS_WORKSPACES."
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    Reason:
      type: string
      enum:
      - CONFLICT
      - MERGE_CONFLICT
      - NOT_FAST_FORWARD
      - ALREADY_INTEGRATED
      - PUSH_REJECTED
      - RELEASE_REQUIRED
...
    Response21:
      type: object
      properties:
        draft:
          $ref: "#/components/schemas/WorkspacePromptDraftDto"
    Response22:
      type: object
      properties:
        draft:
          $ref: "#/components/schemas/WorkspacePromptDraftDto"
...
    RunnerOwnedWorkspace:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The workspace row id
        repositoryId:
          type: string
          description: Its repository
        branch:
          type: string
          description: "Its branch, or null"
...
    SavePromptDraftRequest:
      type: object
      required:
      - content
      properties:
        content:
          type: string
        serializedPrompt:
          type: string
...
    WorkspacePromptDraftDto:
      type: object
      properties:
        content:
          type: string
        serializedPrompt:
          type: string
        promptVersion:
          type: integer
          format: int64
        lastRunAt:
          $ref: "#/components/schemas/Instant"
        lastRunPromptVersion:
          type: integer
          format: int64
        lastRunCommandId:
          type: string
        updatedAt:
          $ref: "#/components/schemas/Instant"
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
...
  /workspaces/api/workspaces/{id}/prompt-draft:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePromptDraftRequest"
        required: true
      responses:
        "200":
          description: "The persisted draft, with its fresh updatedAt."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response21"
        "400":
          description: The content is not a well-formed JSON document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No such ACTIVE workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: The draft exceeds the configured byte cap.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Save
      tags:
      - Workspace Prompt Draft Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: The saved draft.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response22"
        "404":
          description: "No such ACTIVE workspace, or it has no saved draft."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get
      tags:
      - Workspace Prompt Draft Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: There is no draft for this workspace.
        "404":
          description: No such ACTIVE workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Delete
      tags:
      - Workspace Prompt Draft Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
...
info:
  title: qits-workspaces API
  version: 2026.1005.203201